+++ activate.sh
@@ -25,9 +25,9 @@ function tar-diff() {
 
 if [ -d venv ]; then
   source venv/bin/activate
-elif [ -z "$CONDA_PREFIX" ]; then
+elif [[ -z "$CONDA_PREFIX" ]] && [[ -z "$VIRTUAL_ENV" ]]; then
   echo 'WARNING: no venv / conda environment present.'
-  echo 'Read the docs on how to setup your Python environment: https://exabl.github.io/eturb/README.html#easy-way'
+  echo 'Read the docs on how to setup your Python environment: https://exabl.github.io/snek5000/README.html#easy-way'
 fi
 
 if [ "$BASH_VERSION" ]; then