+++ docs/conf.py
@@ -22,9 +22,7 @@ from pathlib import Path
 from subprocess import PIPE
 
 import breathe
-
-import abl
-import eturb
+import snek5000
 
 
 def root(module):
@@ -36,23 +34,22 @@ sys.path.insert(0, root(breathe))
 # If extensions (or modules to document with autodoc) are in another directory,
 # add these directories to sys.path here. If the directory is relative to the
 # documentation root, use os.path.abspath to make it absolute, like shown here.
-sys.path.insert(0, root(abl))
-sys.path.insert(0, root(eturb))
+sys.path.insert(0, root(snek5000))
 
 print("sys.path =\n   ", "\n    ".join(sys.path))
 
 # -- Project information -----------------------------------------------------
 
-project = "eturb"
+project = "snek5000"
 _today = date.today()
 copyright = (
     f"2019 - {_today.year}, Ashwin Vishnu Mohanan. Published: {_today.isoformat()}"
 )
 author = "Ashwin Vishnu Mohanan"
 
-version = ".".join(eturb.__version__.split(".")[:3])
+version = ".".join(snek5000.__version__.split(".")[:3])
 # The full version, including alpha/beta/rc tags
-release = eturb.__version__
+release = snek5000.__version__
 
 
 # -- General configuration ---------------------------------------------------
@@ -74,10 +71,7 @@ extensions = [
 os.makedirs("_build/html/doxygen", exist_ok=True)
 
 # Inspect whether to run doxygen or not
-last_modified = max(
-    eturb.util.last_modified("../lib").stat().st_mtime,
-    eturb.util.last_modified("../src/abl").stat().st_mtime,
-)
+last_modified = snek5000.util.last_modified("../lib").stat().st_mtime
 timestamp = Path("_build/.doxygen_timestamp")
 if timestamp.exists() and Path("_build/xml").exists():
     with open(timestamp) as fp:
@@ -132,11 +126,11 @@ else:
     breathe_domain_by_file_pattern = {"SIZE": "f"}
 
     # Input sources
-    breathe_projects = {"eturb": "_build/xml/"}
+    breathe_projects = {"snek5000": "_build/xml/"}
     #  breathe_projects_source = {
     #      "abl": ("../src/abl", ["SIZE", "abl.usr"]),
     #  }
-    breathe_default_project = "eturb"
+    breathe_default_project = "snek5000"
 
 # ----------------------------------------------------------------------------