+++ src/snek5000/solvers/__init__.py
@@ -6,6 +6,5 @@ Solver framework
 
    base
    kth
-   abl
 
 """